[#2038] - Encapsula la elección de variante en un único punto de entrada
La página conocía las dos variantes y el mapeo del contexto de navegación hacia
cada una: importaba ambos componentes y resolvía la elección en su plantilla. Esa
regla se habría duplicado apenas la página de lectura integrara el bloque.

ReadingSuggestions pasa a ser ese punto de entrada y se queda con el nombre; el
presentacional que lo tenía pasa a ReadingSuggestionsList. La página monta un
solo elemento y solo transporta el contexto.

Cada variante vive en su propio bloque diferido, así se descarga únicamente el
bundle de la que corresponde: el bundle emitido resuelve cada rama con su propia
importación dinámica, de poco menos de 1 KB comprimido cada una.
