+++ src/app/components/reading-suggestions/author-reading-suggestions.component.stories.ts
@@ -26,7 +26,7 @@ const meta: Meta<AuthorReadingSuggestionsComponent> = {
 		docs: {
 			canvas: { sourceState: 'shown' },
 			description: {
-				component: `<div><p>Envoltorio conectado que sugiere otras obras del mismo autor al pie de la lectura. Resuelve las obras contra el proveedor de datos, descarta la que se está leyendo, elige tres al azar y delega toda la presentación en <a href="./?path=/docs/componentes-v3-readingsuggestions--docs" target="_top"><strong>ReadingSuggestions</strong></a>.</p><p>Las tarjetas ocultan avatar y nombre del autor: el encabezado del bloque ya lo nombra. El azar se resuelve una sola vez por consulta, al resolverse los datos, para que las sugerencias no se rebarajen mientras la persona lee.</p><p>El fetch no bloquea el renderizado en el servidor: la página que lo consume monta el bloque dentro de un <code>&#64;defer (on viewport)</code>, así los datos se piden una sola vez y ya en el cliente, al acercarse el final de la lectura.</p></div>`,
+				component: `<div><p>Envoltorio conectado que sugiere otras obras del mismo autor al pie de la lectura. Resuelve las obras contra el proveedor de datos, descarta la que se está leyendo, elige tres al azar y delega toda la presentación en <a href="./?path=/docs/componentes-v3-readingsuggestionslist--docs" target="_top"><strong>ReadingSuggestionsList</strong></a>.</p><p>Las tarjetas ocultan avatar y nombre del autor: el encabezado del bloque ya lo nombra. El azar se resuelve una sola vez por consulta, al resolverse los datos, para que las sugerencias no se rebarajen mientras la persona lee.</p><p>El fetch no bloquea el renderizado en el servidor: la página que lo consume monta el bloque dentro de un <code>&#64;defer (on viewport)</code>, así los datos se piden una sola vez y ya en el cliente, al acercarse el final de la lectura.</p></div>`,
 			},
 		},
 		layout: 'padded',
@@ -92,7 +92,7 @@ export const Cargando: Story = {
 		docs: {
 			description: {
 				story:
-					'Estado de carga del envoltorio, mientras las obras del autor viajan. El alternador real↔esqueleto vive en la story <strong>Estados</strong> de <a href="./?path=/docs/componentes-v3-readingsuggestions--docs" target="_top"><strong>ReadingSuggestions</strong></a>, que es quien renderiza ambos estados.',
+					'Estado de carga del envoltorio, mientras las obras del autor viajan. El alternador real↔esqueleto vive en la story <strong>Estados</strong> de <a href="./?path=/docs/componentes-v3-readingsuggestionslist--docs" target="_top"><strong>ReadingSuggestionsList</strong></a>, que es quien renderiza ambos estados.',
 			},
 		},
 	},
