+++ src/app/pages/story/story.component.ts
@@ -20,8 +20,7 @@ import { StoryStructuredDataDirective } from './story-structured-data.directive'
 import { STORY_HOST, type StoryHost } from './story-host';
 
 // Components
-import { AuthorReadingSuggestionsComponent } from '@components/reading-suggestions/author-reading-suggestions.component';
-import { CollectionReadingSuggestionsComponent } from '@components/reading-suggestions/collection-reading-suggestions.component';
+import { ReadingSuggestionsComponent } from '@components/reading-suggestions/reading-suggestions.component';
 import { BioSummaryCardComponent } from '@components/bio-summary-card/bio-summary-card.component';
 import { ShareContentComponent } from '@components/share-content/share-content.component';
 import { EditorialTextBlockComponent } from '@components/editorial-text-block/editorial-text-block.component';
@@ -36,9 +35,7 @@ import { faSolidArrowRightLong } from '@ng-icons/font-awesome/solid';
 	selector: 'cuentoneta-story',
 	templateUrl: './story.component.html',
 	imports: [
-		AuthorReadingSuggestionsComponent,
 		BioSummaryCardComponent,
-		CollectionReadingSuggestionsComponent,
 		CommonModule,
 		EditorialTextBlockComponent,
 		MediaResourceComponent,
@@ -47,6 +44,7 @@ import { faSolidArrowRightLong } from '@ng-icons/font-awesome/solid';
 		RouterLink,
 		ShareContentComponent,
 		ProgressBarComponent,
+		ReadingSuggestionsComponent,
 		NgIcon,
 	],
 	providers: [
